initial change to test with CI for CPU/CUDA errors
