+++ source/slang/slang-ir-legalize-varying-params.h
@@ -6,12 +6,12 @@ namespace Slang
 
 class DiagnosticSink;
 
-struct IRFunc;
 struct IRModule;
 struct IRInst;
 struct IRFunc;
 struct IRVectorType;
 struct IRBuilder;
+struct IREntryPointDecoration;
 
 void legalizeEntryPointVaryingParamsForCPU(
     IRModule*               module,
@@ -21,6 +21,17 @@ void legalizeEntryPointVaryingParamsForCUDA(
     IRModule*               module,
     DiagnosticSink*         sink);
 
+void legalizeEntryPointVaryingParamsForMetal(
+    IRModule*               module,
+    DiagnosticSink*         sink);
+
+struct EntryPointInfo
+{
+    IRFunc* entryPointFunc;
+    IREntryPointDecoration* entryPointDecor;
+};
+
+//TODO: remove these below
 IRInst* emitCalcGroupThreadIndex(
     IRBuilder& builder,
     IRInst* groupThreadID,
